feat: implementation of transducer class
Implementation of the Transducer class, which is responsible for
applying transduction rules. Additionally, a TransducerBuilder was
created to simplify the construction loigic of a Transducer, since it
receives one strategy to define the iterator transversal order and
another strategy to determine the next node to analyze after a
modification occurs in the SyntaxTree.